third_party/WebKit/LayoutTests/platform/win/http/tests/cookies/same-site/popup-cross-site-expected.txt

This is a testharness.js-based test.
PASS 'example.test' is not same-site with 'cookies.not-example.test', so strict samesite cookies are not sent. 
Harness: the test ran to completion.
